OpenFaaS Client restored
