+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod consts;
 pub mod crds;
+mod openfaas_client;
 pub mod operator;
-pub mod util;
