+++ src/openfaas_client/mod.rs
@@ -1,0 +1,3 @@
+#[allow(dead_code)]
+pub mod client;
+mod request;
