+++ src/openfaas_client/request/mod.rs
@@ -1,0 +1,1 @@
+pub mod functions;
